+++ NEWS.md
@@ -1,5 +1,5 @@
 ## Version 0.2.2
 
-* Moved CIC calculation in C++.
+* Moved CIC calculation into C++.
 * Added testing.
-
+* Added influence function calculation.